services/auth_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
asyncpg
pydantic
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart